--- test/uport-connect.js
@@ -1,2 +1,0 @@
-import Connect from '../src/Connect.js'
-export { Connect }
